Split joystick_get to definition/declaration
