+++ mos-platform/cx16/CMakeLists.txt
@@ -58,6 +58,7 @@ add_platform_library(cx16-c
  cx16_k_graph_set_window.s
  cx16_k_i2c_read_byte.s
  cx16_k_i2c_write_byte.s
+ cx16_k_joystick_get.c
  cx16_k_joystick_scan.s
  cx16_k_kbdbuf_get_modifiers.s
  cx16_k_kbdbuf_peek.s
@@ -98,3 +99,4 @@ add_platform_library(cx16-c
 target_include_directories(cx16-c BEFORE PUBLIC .)
 target_compile_options(cx16-c PUBLIC -mcpu=mosw65c02)
 target_link_libraries(cx16-c PRIVATE common-asminc)
+target_compile_definitions(cx16-c PUBLIC __CX16__)
